+++ ztb/training/unified_trainer/reporting.py
@@ -13,7 +13,7 @@ from ztb.reporting.services.training_reports import (
     save_ensemble_report,
     save_training_report,
 )
-from ztb.training.utils.training_stats_payloads import extract_reward_component_metrics
+from ztb.training.utils.training_stats_payloads import attach_reward_component_metrics
 from ztb.types.common import ConfigDict, ObjectMap, ObjectRecords
 from ztb.utils.logging_utils import get_logger
 from ztb.utils.time_utils import current_iso_timestamp
@@ -55,9 +55,7 @@ class TrainingReporter:
             report["training_events"] = self.events.copy()
 
         # Include reward_components if present in stats for AB analysis
-        reward_components = extract_reward_component_metrics(stats)
-        if reward_components is not None:
-            report["reward_components"] = reward_components
+        attach_reward_component_metrics(report, stats)
 
         return report